requirements.txt
aiogram>=3.27.0
telethon>=1.43.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
python-dateutil>=2.9.0
aiohttp>=3.10.0
